coap: Support block-wise transfer with non-piggybacked server transactions
Wakaama uses the transcations mechanism to generate follow up block
request for block1 transfers. The transaction contains the full block
context and payload. This works only for piggybacked transactions
because if the client sends an empty ack to answer a POST or PUT request
the transaction is finished.

This fix tracks request block1 transfers separately to reconstruct the
context and payload also for CON continuous 2.31 messages and allows
non-piggybacked block1 block-wise transfers.

Not tested: block2 and the client side.