coap: Support block-wise transfers with non-piggybacked transactions
Transactions are the core machanism to handle the CoAP request/response
mechanics. Wakaama uses the transactions to generate follow up block
requests for block1 transfers. A transaction contains the full block
context and payload for consequtive transmissions.

This fix uses the available transaction created for the block1 request
to create the next request also for non-piggybacked replies. In this
case, the transaction token is used to get the information from the
previous block.

Not tested: block2 and the client side.